move user authentication view components
