+++ client/src/routes.ts
@@ -5,12 +5,12 @@ import OurStory from './views/OurStory';
 import TrustAndSafety from './views/TrustAndSafety';
 import TermsOfService from './views/TermsOfService';
 import PrivacyPolicy from './views/PrivacyPolicy';
-import Login from './components/Users/Auth/Login';
-import Signup from './components/Users/Auth/Signup';
-import SignupCitizen from './components/Users/Auth/SignupCitizen';
-import SignupNonProfit from './components/Users/Auth/SignupNonProfit';
-import Inbox from './components/Users/Inbox/Inbox';
-import User from './components/Users/User';
+import Login from './views/Login';
+import Signup from './views/Signup';
+import SignupCitizen from './views/SignupCitizen';
+import SignupNonProfit from './views/SignupNonProfit';
+import Inbox from './views/Inbox';
+import User from './views/User';
 import ActionForm from './views/ActionForm';
 import AssetsView from './views/AssetsView';
 import Asset from './views/Asset';
