+++ client/src/views/Signup.tsx
@@ -6,7 +6,7 @@ import { Grid, Container, Button } from '@mui/material';
 
 import type { Theme } from '@mui/material/styles';
 
-import routes from '../../../routes';
+import routes from '../routes';
 
 const useStyles = makeStyles((theme: Theme) => ({
   imageBackground: {
